group io get code
